+++ liqwid-plutarch-extra/src/Plutarch/Extra/TermCont.hs
@@ -1,8 +1,12 @@
 {-# LANGUAGE PolyKinds #-}
 
 module Plutarch.Extra.TermCont (
+  -- * Conditional guards
   pguardWithC,
   pguardShowC,
+  -- * Marker logging
+  markInfoC,
+  markDebugC
 ) where
 
 {- | 'pguardC' but with type threading for better traces.
@@ -47,3 +51,39 @@ pguardShowC ::
   TermCont @r s ()
 pguardShowC message =
   pguardWithC (\t -> message <> " Guarded object was: " <> pshow t)
+
+-- | Given an action, a pre-action marker and a post-action marker, log the
+-- pre-action marker, do the action, then log the post-action marker. All
+-- logging will be done in info mode.
+--
+-- This runs in 'TermCont' to ensure sensible sequencing.
+--
+-- @since 4.0.1
+markInfoC :: forall (r :: S -> Type) (a :: S -> Type) (s :: S) . 
+  -- | Pre-action marker
+  Term s PString -> 
+  -- | Post-action marker
+  Term s PString -> 
+  TermCont @r s (Term s a) -> 
+  TermCont @r s (Term s a)
+markInfoC pre post act = do
+  tcont $ \f -> ptraceInfo pre (f ())
+  x <- act
+  tcont $ \f -> ptraceInfo post (f ())
+  pure x
+
+-- | As 'markInfoC', but using debug tracing mode instead.
+--
+-- @since 4.0.1
+markDebugC :: forall (r :: S -> Type) (a :: S -> Type) (s :: S) . 
+  -- | Pre-action marker
+  Term s PString -> 
+  -- | Post-action marker
+  Term s PString -> 
+  TermCont @r s (Term s a) -> 
+  TermCont @r s (Term s a)
+markDebugC pre post act = do
+  tcont $ \f -> ptraceDebug pre (f ())
+  x <- act
+  tcont $ \f -> ptraceDebug post (f ())
+  pure x